:sparkles: Move bitCastOr
